Move Ukraine data loader into its own module
It's not a generic Popolo loader any more so let's not pretend it is.
